Refactoring User Data Transfer Objects
