+++ src/main/java/com/unsa/suppliers/domain/dtos/users/UserRequest.java
@@ -1,5 +1,6 @@
-package com.unsa.suppliers.domain.dtos;
+package com.unsa.suppliers.domain.dtos.users;
 
+import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -10,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Data @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserUpdateRequest {
+public class UserRequest {
     @NotBlank(message = "Name must not be blank")
     @Size(min = 3, message = "Name must contain at least 3 characters")
     private String name;
@@ -20,4 +21,10 @@ public class UserUpdateRequest {
     @NotBlank(message = "Username must not be blank")
     @Size(min = 4, message = "Username must contain at least 4 characters")
     private String username;
+    @NotBlank(message = "Email must not be blank")
+    @Email(message = "Please send a valid email")
+    private String email;
+    @NotBlank(message = "Password must not be blank")
+    @Size(min = 8, message = "Password must contain at least 8 characters")
+    private String password;
 }
