+++ src/main/java/com/unsa/suppliers/domain/dtos/users/UserUpdateRequest.java
@@ -1,6 +1,5 @@
-package com.unsa.suppliers.domain.dtos;
+package com.unsa.suppliers.domain.dtos.users;
 
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -11,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Data @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserRequest {
+public class UserUpdateRequest {
     @NotBlank(message = "Name must not be blank")
     @Size(min = 3, message = "Name must contain at least 3 characters")
     private String name;
@@ -21,10 +20,4 @@ public class UserRequest {
     @NotBlank(message = "Username must not be blank")
     @Size(min = 4, message = "Username must contain at least 4 characters")
     private String username;
-    @NotBlank(message = "Email must not be blank")
-    @Email(message = "Please send a valid email")
-    private String email;
-    @NotBlank(message = "Password must not be blank")
-    @Size(min = 8, message = "Password must contain at least 8 characters")
-    private String password;
 }
